+++ packages/core/src/__tests__/renderLeafPlugins/without-plugins.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { render } from '@testing-library/react';
-import { renderLeafPlugins } from '../../utils';
+import { renderLeafPlugins } from '../../utils/renderLeafPlugins';
 
 const attributes = {
   'data-testid': 'Leaf',
