+++ packages/slate-plugins/src/__tests__/deserializers/deserialize-html/withDeserializeHtml/iframe-without-src.spec.tsx
@@ -1,11 +1,11 @@
 /** @jsx jsx */
 
-import { jsx } from '__test-utils__/jsx';
-import { withDeserializeHTML } from 'deserializers/deserialize-html';
-import { MediaEmbedPlugin } from 'elements/media-embed';
 import { Editor } from 'slate';
 import { withReact } from 'slate-react';
+import { jsx } from '../../../../__test-utils__/jsx';
+import { withDeserializeHTML } from '../../../../deserializers/deserialize-html';
 import { withInlineVoid } from '../../../../element';
+import { MediaEmbedPlugin } from '../../../../elements/media-embed';
 
 const input = ((
   <editor>
