+++ packages/slate-plugins/src/__tests__/deserializers/deserialize-html/withDeserializeHtml/plain-text.spec.tsx
@@ -1,9 +1,9 @@
 /** @jsx jsx */
 
-import { jsx } from '__test-utils__/jsx';
-import { withDeserializeHTML } from 'deserializers/deserialize-html';
 import { Editor } from 'slate';
 import { withReact } from 'slate-react';
+import { jsx } from '../../../../__test-utils__/jsx';
+import { withDeserializeHTML } from '../../../../deserializers/deserialize-html';
 
 const input = ((
   <editor>
